tidy and abstract synchronisation tests
